+++ packages/article-image/article-image.stories.js
@@ -1,7 +1,6 @@
 import "react-native";
 import React from "react";
 import { storiesOf } from "@storybook/react-native";
-import { checkA11y } from "@storybook/addon-a11y";
 import ArticleImage from "./article-image";
 
 const primaryImage = require("./fixtures/primary-image.json").fixture;
@@ -12,7 +11,6 @@ const portraitInlineImage = require("./fixtures/portrait-inline-image.json")
   .fixture;
 
 storiesOf("Primitives/ArticleImage", module)
-  .addDecorator(checkA11y)
   .add("Primary", () => (
     <ArticleImage
       imageOptions={primaryImage.imageOptions}