+++ packages/gestures/gestures.stories.native.js
@@ -2,7 +2,6 @@ import React from "react";
 import { View, Text, StyleSheet } from "react-native";
 import { storiesOf } from "@storybook/react-native";
 import { CenteredDecorator } from "@times-components/storybook";
-import { checkA11y } from "@storybook/addon-a11y";
 import Gestures from "./gestures";
 
 const styles = StyleSheet.create({
@@ -40,7 +39,6 @@ const styles = StyleSheet.create({
 });
 
 storiesOf("Helpers/Gestures", module)
-  .addDecorator(checkA11y)
   .addDecorator(CenteredDecorator)
   .add("With a View", () => (
     <Gestures style={styles.gestures}>