+++ packages/gradient/gradient.stories.js
@@ -2,7 +2,6 @@ import React from "react";
 import PropTypes from "prop-types";
 import { StyleSheet, View } from "react-native";
 import { storiesOf } from "@storybook/react-native";
-import { checkA11y } from "@storybook/addon-a11y";
 import Gradient from "./gradient";
 
 const styles = StyleSheet.create({
@@ -32,7 +31,6 @@ Component.propTypes = {
 };
 
 storiesOf("Primitives/Gradient", module)
-  .addDecorator(checkA11y)
   .add("Default", () => (
     <Component
       style={{