+++ packages/tracking/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.9.2"></a>
+## [0.9.2](https://github.com/newsuk/times-components/compare/@times-components/tracking@0.9.1...@times-components/tracking@0.9.2) (2018-03-06)
+
+
+
+
+**Note:** Version bump only for package @times-components/tracking
+
+<a name="0.9.1"></a>
+## [0.9.1](https://github.com/newsuk/times-components/compare/@times-components/tracking@0.9.0...@times-components/tracking@0.9.1) (2018-03-02)
+
+
+
+
+**Note:** Version bump only for package @times-components/tracking
+
 <a name="0.9.0"></a>
 # [0.9.0](https://github.com/newsuk/times-components/compare/@times-components/tracking@0.7.8...@times-components/tracking@0.9.0) (2018-03-01)